Search for author by ID

+++ src/main/java/com/ciaranmckenna/readinglistapp/dao/repository/AuthorRepository.java
@@ -3,5 +3,5 @@ package com.ciaranmckenna.readinglistapp.dao.repository;
 import com.ciaranmckenna.readinglistapp.dao.entity.Author;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AuthorRepository extends JpaRepository<Author, Long> {
+public interface AuthorRepository extends JpaRepository<Author, Integer> {
 }
